+++ admin/compiler/joomla_3/Helper_site.php
@@ -17,12 +17,7 @@ defined('_JEXEC') or die('Restricted access');
 // No direct access to this file
 defined('_JEXEC') or die('Restricted access');
 
-// The power autoloader for this project site area.
-$power_autoloader = JPATH_SITE . '/components/com_###component###/helpers/powerloader.php';
-if (file_exists($power_autoloader))
-{
-	require_once $power_autoloader;
-}
+###SITE_CUSTOM_POWER_AUTOLOADER###
 
 ###SITE_HELPER_CLASS_HEADER###