+++ admin/compiler/joomla_3/component_admin.php
@@ -17,12 +17,7 @@ defined('_JEXEC') or die('Restricted access');
 // No direct access to this file
 defined('_JEXEC') or die('Restricted access');
 
-// The power autoloader for this project admin area.
-$power_autoloader = JPATH_ADMINISTRATOR . '/components/com_###component###/helpers/powerloader.php';
-if (file_exists($power_autoloader))
-{
-	require_once $power_autoloader;
-}
+###CUSTOM_POWER_AUTOLOADER###
 
 ###ADMIN_COMPONENT_HEADER###