+++ admin/controllers/compiler.php
@@ -22,6 +22,7 @@ use VDM\Joomla\Componentbuilder\Compiler\Factory as CFactory;
 use Joomla\CMS\Version;
 use VDM\Joomla\Utilities\ArrayHelper as UtilitiesArrayHelper;
 use VDM\Joomla\Utilities\StringHelper;
+use Joomla\CMS\Uri\Uri;
 
 /**
  * Compiler Admin Controller
@@ -127,7 +128,7 @@ class ComponentbuilderControllerCompiler extends AdminController
 				$redirect_url = Route::_('index.php?option=com_componentbuilder&view=compiler', false);
 				if (($pos = strpos($model->compiler->filepath['component'], "/tmp/")) !== FALSE)
 				{
-				    $url = \JUri::root() . substr($model->compiler->filepath['component'], $pos + 1);
+				    $url = Uri::root() . substr($model->compiler->filepath['component'], $pos + 1);
 				}
 				// check if we have plugins
 				$add_plugin_install = UtilitiesArrayHelper::check($model->compiler->filepath['plugins'], true);
@@ -210,7 +211,7 @@ class ComponentbuilderControllerCompiler extends AdminController
 							$message[] = '<b>Module Path:</b> <code>' . $module_path . '</code><br />';
 							if (($pos = strpos($module_path, "/tmp/")) !== FALSE)
 							{
-								$module_urls[$module_id] = \JUri::root() . substr($module_path, $pos + 1);
+								$module_urls[$module_id] = Uri::root() . substr($module_path, $pos + 1);
 								$message[] = '<b>Module URL:</b> <code>' . $module_urls[$module_id] . '</code><br />';
 							}
 						}
@@ -226,7 +227,7 @@ class ComponentbuilderControllerCompiler extends AdminController
 							$message[] = '<b>Plugin Path:</b> <code>' . $plugin_path . '</code><br />';
 							if (($pos = strpos($plugin_path, "/tmp/")) !== FALSE)
 							{
-								$plugin_urls[$plugin_id] = \JUri::root() . substr($plugin_path, $pos + 1);
+								$plugin_urls[$plugin_id] = Uri::root() . substr($plugin_path, $pos + 1);
 								$message[] = '<b>Plugin URL:</b> <code>' . $plugin_urls[$plugin_id] . '</code><br />';
 							}
 						}
@@ -532,7 +533,7 @@ class ComponentbuilderControllerCompiler extends AdminController
 		// check if user has the right
 		$user = Factory::getUser();
 		// set page redirect
-		$redirect_url = \JRoute::_('index.php?option=com_componentbuilder&view=compiler', false);
+		$redirect_url = Route::_('index.php?option=com_componentbuilder&view=compiler', false);
 		// set massage
 		$message = Text::_('COM_COMPONENTBUILDER_YOU_DO_NOT_HAVE_PERMISSION_TO_RUN_THE_EXPANSION_MODULE');
 		// check if this user has the right to run expansion
@@ -541,7 +542,7 @@ class ComponentbuilderControllerCompiler extends AdminController
 			// set massage
 			$message = Text::_('COM_COMPONENTBUILDER_EXPANSION_FAILED_PLEASE_CHECK_YOUR_SETTINGS_IN_THE_GLOBAL_OPTIONS_OF_JCB_UNDER_THE_DEVELOPMENT_METHOD_TAB');
 			// run expansion via API
-			$result = ComponentbuilderHelper::getFileContents(\JUri::root() . 'index.php?option=com_componentbuilder&task=api.expand');
+			$result = ComponentbuilderHelper::getFileContents(Uri::root() . 'index.php?option=com_componentbuilder&task=api.expand');
 			// is there a message returned
 			if (!is_numeric($result) && StringHelper::check($result))
 			{