+++ admin/views/admin_view/view.html.php
@@ -217,7 +217,7 @@ class ComponentbuilderViewAdmin_view extends HtmlView
 		// add var key
 		$this->document->addScriptDeclaration("var vastDevMod = '" . $this->get('VDM') . "';");
 		// add return_here
-		$this->document->addScriptDeclaration("var return_here = '" . urlencode(base64_encode((string) \JUri::getInstance())) . "';");
+		$this->document->addScriptDeclaration("var return_here = '" . urlencode(base64_encode((string) Uri::getInstance())) . "';");
 		Text::script('view not acceptable. Error');
 	}