+++ admin/views/get_snippets/tmpl/default.php
@@ -36,7 +36,7 @@ Html::_('behavior.keepalive');
 	}
 </script>
 
-<form action="<?php echo \JRoute::_('index.php?option=com_componentbuilder'); ?>" method="post" name="adminForm" id="adminForm" class="form-validate" enctype="multipart/form-data">
+<form action="<?php echo Route::_('index.php?option=com_componentbuilder'); ?>" method="post" name="adminForm" id="adminForm" class="form-validate" enctype="multipart/form-data">
 	<input type="hidden" name="task" value="" />
 	<?php echo Html::_('form.token'); ?>
 </form>