+++ admin/views/libraries/tmpl/default_body.php
@@ -17,6 +17,7 @@ use Joomla\CMS\Language\Text;
 use Joomla\CMS\HTML\HTMLHelper as Html;
 use VDM\Joomla\Utilities\ArrayHelper;
 use VDM\Joomla\Utilities\GetHelper;
+use Joomla\CMS\Uri\Uri;
 
 $edit = "index.php?option=com_componentbuilder&view=libraries&task=library.edit";
 
@@ -77,7 +78,7 @@ $edit = "index.php?option=com_componentbuilder&view=libraries&task=library.edit"
 			<?php	// setup the return path
 				if (!isset($returnpath))
 				{
-					$returnpath = urlencode(base64_encode((string) \JUri::getInstance()));
+					$returnpath = urlencode(base64_encode((string) Uri::getInstance()));
 				}
 				// setup the buttons
 				if (!isset($_buttons) || !ArrayHelper::check($_buttons))