+++ site/helpers/componentbuilder.php
@@ -12,7 +12,7 @@
 // No direct access to this file
 defined('_JEXEC') or die('Restricted access');
 
-// The power autoloader for this project site area.
+// The power autoloader for this project (JPATH_SITE) area.
 $power_autoloader = JPATH_SITE . '/components/com_componentbuilder/helpers/powerloader.php';
 if (file_exists($power_autoloader))
 {
@@ -58,6 +58,7 @@ use VDM\Joomla\Utilities\String\PluginHelper;
 use VDM\Joomla\Utilities\GuidHelper;
 use VDM\Joomla\Utilities\Component\Helper;
 use VDM\Joomla\Utilities\FormHelper;
+use Joomla\CMS\Router\Route;
 
 /**
  * Componentbuilder component helper
@@ -2748,7 +2749,7 @@ abstract class ComponentbuilderHelper
 				// get the global settings
 				if (!ObjectHelper::check(self::$params))
 				{
-					self::$params = \JComponentHelper::getParams('com_componentbuilder');
+					self::$params = ComponentHelper::getParams('com_componentbuilder');
 				}
 				self::$gitHubAccessToken = self::$params->get('github_access_token', null);
 			}
@@ -3251,7 +3252,7 @@ abstract class ComponentbuilderHelper
 			$script['view'][] = PHP_EOL . "</script>";
 			$script['view'][] = "";
 			$script['view'][] = PHP_EOL . "<div id=\"installer-import\" class=\"clearfix\">";
-			$script['view'][] = "<form enctype=\"multipart/form-data\" action=\"<?php echo \JRoute::_('index.php?option=com_[[[-#-#-component]]]&view=import_[[[-#-#-views]]]');?>\" method=\"post\" name=\"adminForm\" id=\"adminForm\" class=\"form-horizontal form-validate\">";
+			$script['view'][] = "<form enctype=\"multipart/form-data\" action=\"<?php echo Route::_('index.php?option=com_[[[-#-#-component]]]&view=import_[[[-#-#-views]]]');?>\" method=\"post\" name=\"adminForm\" id=\"adminForm\" class=\"form-horizontal form-validate\">";
 			$script['view'][] = "";
 			$script['view'][] = PHP_EOL . self::_t(1) . "<?php if (!empty( \$this->sidebar)) : ?>";
 			$script['view'][] = self::_t(2) . "<div id=\"j-sidebar-container\" class=\"span2\">";
@@ -3654,7 +3655,7 @@ abstract class ComponentbuilderHelper
 		// get the global settings
 		if (!ObjectHelper::check(self::$params))
 		{
-			self::$params = \JComponentHelper::getParams('com_componentbuilder');
+			self::$params = ComponentHelper::getParams('com_componentbuilder');
 		}
 		$folderPath = self::$params->get($target, $default);
 		// create the folder if it does not exist
@@ -4883,7 +4884,7 @@ abstract class ComponentbuilderHelper
 			// get the global settings
 			if (!ObjectHelper::check(self::$params))
 			{
-				self::$params = \JComponentHelper::getParams('com_componentbuilder');
+				self::$params = ComponentHelper::getParams('com_componentbuilder');
 			}
 			// get UIKIT version
 			$uikit = self::$params->get('uikit_version', 2);
@@ -4995,7 +4996,7 @@ abstract class ComponentbuilderHelper
 			// get the global settings
 			if (!ObjectHelper::check(self::$params))
 			{
-				self::$params = \JComponentHelper::getParams('com_componentbuilder');
+				self::$params = ComponentHelper::getParams('com_componentbuilder');
 			}
 			// get UIKIT version
 			$uikit = self::$params->get('uikit_version', 2);
@@ -5101,7 +5102,7 @@ abstract class ComponentbuilderHelper
 			// set the edit link
 			if ($jRoute)
 			{
-				return \JRoute::_("index.php?option=" . $component . "&view=" . $views . "&task=" . $view . ".edit&id=" . $record->id . $ref);
+				return Route::_("index.php?option=" . $component . "&view=" . $views . "&task=" . $view . ".edit&id=" . $record->id . $ref);
 			}
 			return "index.php?option=" . $component . "&view=" . $views . "&task=" . $view . ".edit&id=" . $record->id . $ref;
 		}